refactor: add lib/utils/managed.ts
