+++ packages/waku/src/lib/utils/managed.ts
@@ -1,0 +1,32 @@
+import { EXTENSIONS } from '../builder/constants.js';
+
+export const getManagedServerEntry = (config: {
+  srcDir: string;
+  pagesDir: string;
+  apiDir: string;
+}) => {
+  const globBase = `/${config.srcDir}/${config.pagesDir}/`;
+  const globPattern = `${globBase}**/*.{${EXTENSIONS.map((ext) => ext.slice(1)).join(',')}}`;
+  return `
+import { unstable_fsRouter as fsRouter } from 'waku/router/server';
+const glob = import.meta.glob(${JSON.stringify(globPattern)}, { base: ${JSON.stringify(globBase)} });
+const apiDir = ${JSON.stringify(config.apiDir)};
+export default fsRouter(glob, { apiDir });
+`;
+};
+
+export const getManagedClientEntry = () => {
+  return `
+import { StrictMode, createElement } from 'react';
+import { createRoot, hydrateRoot } from 'react-dom/client';
+import { Router } from 'waku/router/client';
+
+const rootElement = createElement(StrictMode, null, createElement(Router));
+
+if (globalThis.__WAKU_HYDRATE__) {
+  hydrateRoot(document, rootElement);
+} else {
+  createRoot(document).render(rootElement);
+}
+`;
+};
